refactor: remove agent-prompt.ts file
